Fix compiler warnings for C23 compatibility
- Fix typedef union PCode to include type name in cgen.h
- Add function prototypes for rindex() and badFile() in xbitmap.c
- Fix SlotStruct pointer type in objs.c array initialization
- Add function prototypes for HashTable->get and HashTable->remove methods
- Add function prototypes for ClassInfo->slotGetMeth and slotSetMeth
- Fix validObjectP macro to cast pointer to long for HashTable operations
- Fix all HashTable method calls to explicitly cast pointers to long
- Change slotGetMeth/slotSetMeth prototypes from void* to long for correct type matching

+++ src/viola/cl_field.c
@@ -956,7 +956,7 @@ long meth_field_freeSelf(VObj* self, Packet* result, int argc, Packet argv[]) {
             */
 
             SET_window(self, 0); /*insurance*/
-            window2Obj->remove(window2Obj, GET_window(self));
+            window2Obj->remove(window2Obj, (long)GET_window(self));
         }
     }
     meth_generic_freeSelf(self, result, argc, argv);
